Split out graph Expr code from Dataframe Expr code
Closes #142
Supercedes #158
